+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 1.0.0
 - Flutter_dialogs and flutter_bottom_alert_dialogs combined into one package
+- Add default padding to the bottom alert dialog that is the size of the safe area of a device. You can disable this by setting `useSafeAreaPadding` to false.
 
 ## 0.0.2